⚙️ Refactor class components into functional components
