+++ src/App/App.js
@@ -4,10 +4,9 @@ import {
   Route, 
   // Redirect
 } from 'react-router-dom';
-// import { Link } from 'react-router-dom';
 import {LoginForm} from '../LoginForm/LoginForm'
 import { Homepage } from '../Homepage/Homepage'
-import { ErrorBoundary } from '../ErrorMessage/ErrorMessage.js';
+import { ErrorBoundary }  from '../ErrorBoundary/ErrorBoundary';
 import MoviePage from '../MoviePage/MoviePage'
 import { NavBar } from '../NavBar/NavBar'
 import fetcher from '../API/APIcalls';
@@ -70,6 +69,18 @@ export class App extends Component{
     this.setState({movies: promise.movies})
   }
 
+  sortMovies = (value) => {
+    if (!value || value === '--none--') {
+      return false
+    } else if (value === 'descending'){
+      const sortedMovies = this.state.movies.sort((a, b) => a.average_rating - b.average_rating)
+      this.setState({ movies: sortedMovies })
+    } else if (value === 'ascending'){
+      const sortedMovies = this.state.movies.sort((a, b) => b.average_rating - a.average_rating)
+      this.setState({ movies: sortedMovies })
+    } 
+  }
+
   render(){
     return (
       <>
@@ -99,6 +110,7 @@ export class App extends Component{
                   name={this.state.userData}
                   movies={this.state.movies}
                   getMovieDetails={this.getMovieDetails}
+                  sortMovies={this.sortMovies}
                 />
               )
             }}
@@ -113,7 +125,9 @@ export class App extends Component{
               /> 
             }}
           />
-            <Route path='*' render={() => {
+            <Route 
+              exact path='*' 
+              render={() => {
               return  <ErrorBoundary /> //errorMessageData={this.state.hasError}
             }} />
             {/* the above path has a ~50ms 'setTimeout where it will display the error message and then immediately reroute to the correct page */}
