+++ src/ErrorBoundary/ErrorBoundary.scss
