+++ src/Homepage/Homepage.js
@@ -1,23 +1,20 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { WelcomingSection } from '../WelcomingSection/WelcomingSection'
 import Cards  from '../Cards/Cards'
-// import { NavBar } from '../NavBar/NavBar'
 import './Homepage.scss';
 
-export class Homepage extends Component {
-    constructor(props){
-        super(props)
-    }
 
-    render(){
-        return(
-        <>
-          <WelcomingSection name={this.props.name}/>
-          <Cards 
-           getMovieDetails={this.props.getMovieDetails}
-           movies={this.props.movies}
-          />
-        </>
-    )
-  }
+export function Homepage(props) {
+  return(
+    <>
+      <WelcomingSection name={props.name}/>
+      <Cards 
+        movies={props.movies}
+        getMovieDetails={props.getMovieDetails}
+        sortMovies={props.sortMovies}
+      />
+    </>
+)
 }
+
+
